extracted board generation to its own utility file
